==> Faithfulness/4_Faithfulness_Assessment.ipynb <==
import pandas as pd
import ast
import re
import random

# Import generated explanations

# import file
filename = 'NLEs_FaithfulnessStudy.csv'
df = pd.read_csv(filename)

def clean_words(word_set):
    cleaned_set = set(re.sub(r"[^a-zA-Z0-9]", "", word.lower()) for word in word_set)
    return cleaned_set

def calculate_metrics_and_accuracy(row):
    # Preparing the sets
    NLE_words_rationale = clean_words(set(row['NLE_with_rationale'].lower().split())) # Comparison set
    NLE_words_no_rationale = clean_words(set(row['NLE_no_rationale'].lower().split())) # Comparison set
    
    words_list = ast.literal_eval(row['pred_rationales_words'])
    words_list = clean_words(set(word.lower() for word in words_list)) # Ground Truth 
    

    # For rationale
    TP_rationale = len(words_list.intersection(NLE_words_rationale))
    FP_rationale = len(NLE_words_rationale) - TP_rationale
    FN_rationale = len(words_list) - TP_rationale

    
    precision_rationale = TP_rationale / (TP_rationale + FP_rationale) if (TP_rationale + FP_rationale) > 0 else 0
    recall_rationale = TP_rationale / (TP_rationale + FN_rationale) if (TP_rationale + FN_rationale) > 0 else 0
    F1_rationale = (2 * precision_rationale * recall_rationale) / (precision_rationale + recall_rationale) if (precision_rationale + recall_rationale) > 0 else 0
    accuracy_rationale = TP_rationale / (TP_rationale + FP_rationale + FN_rationale) if (TP_rationale + FP_rationale + FN_rationale) > 0 else 0
    
    # For no rationale
    TP_no_rationale = len(words_list.intersection(NLE_words_no_rationale))
    FP_no_rationale = len(NLE_words_no_rationale) - TP_no_rationale
    FN_no_rationale = len(words_list) - TP_no_rationale
    
    precision_no_rationale = TP_no_rationale / (TP_no_rationale + FP_no_rationale) if (TP_no_rationale + FP_no_rationale) > 0 else 0



    recall_no_rationale = TP_no_rationale / (TP_no_rationale + FN_no_rationale) if (TP_no_rationale + FN_no_rationale) > 0 else 0
    F1_no_rationale = (2 * precision_no_rationale * recall_no_rationale) / (precision_no_rationale + recall_no_rationale) if (precision_no_rationale + recall_no_rationale) > 0 else 0
    accuracy_no_rationale = TP_no_rationale / (TP_no_rationale + FP_no_rationale + FN_no_rationale) if (TP_no_rationale + FP_no_rationale + FN_no_rationale) > 0 else 0
    
    return precision_rationale, recall_rationale, F1_rationale, accuracy_rationale, precision_no_rationale, recall_no_rationale, F1_no_rationale, accuracy_no_rationale

# Applying the function and assigning the results to new columns
results = df.apply(calculate_metrics_and_accuracy, axis=1)
df[['precision_rationale', 'recall_rationale', 'F1_rationale', 'accuracy_rationale', 'precision_no_rationale', 'recall_no_rationale', 'F1_no_rationale', 'accuracy_no_rationale']] = pd.DataFrame(results.tolist(), index=df.index)


calculate_mean = ['precision_rationale', 'recall_rationale', 'F1_rationale', 'accuracy_rationale', 'precision_no_rationale', 'recall_no_rationale', 'F1_no_rationale', 'accuracy_no_rationale']

for metric in calculate_mean:
    average_metric = df[str(metric)].mean()
    std_metric = df[str(metric)].std()
    print(f'The overall average for {metric} is: {average_metric} (Std={std_metric})')
    print('___')

# Qualitative Control of Faithfulness

filename = 'NLEs_FaithfulnessStudy.csv'
df = pd.read_csv(filename)

# Define a function to select the rationale column based on the prediction for easier comparison
def select_rationale(row):
    if row['pred'] == 'politics':
        return row['formatted_politics_rationales']
    elif row['pred'] == 'science':
        return row['formatted_science_rationales']
    else:
        return row['formatted_leisure_rationales']

# Apply the function to create a new column
df['selected_rationale'] = df.apply(select_rationale, axis=1)

# Select only the required columns
filtered_df = df[['NLE_with_rationale', 'NLE_no_rationale', 'selected_rationale']]

# If you also need to randomly select 50 instances from this filtered DataFrame:
sampled_filtered_df = filtered_df.sample(n=50, random_state=1)

sampled_filtered_df.to_csv('Qualitative_Assessment.csv', index=False)
